Add 'Time till key' statistic
Adds a new TIME class of tracking statistic that tells you how long
until you reach a certain event. In this case, that event is the
moment you can purchase a key to leave Commerceville.

It uses a 'compound interest with contributions' formula, solved
for the number of loops required until the correcty amount of gold
can be withdrawn as interest from investments.